rename LmdbDirStoreManager package to coop.rchain.store
