+++ casper/src/main/scala/coop/rchain/casper/storage/RNodeKeyValueStoreManager.scala
@@ -3,11 +3,9 @@ package coop.rchain.casper.storage
 import java.nio.file.Path
 
 import cats.effect.Concurrent
-import cats.effect.concurrent.{Deferred, Ref}
-import cats.syntax.all._
 import coop.rchain.shared.Log
-import coop.rchain.shared.store.LmdbDirStoreManager._
-import coop.rchain.shared.store.LmdbDirStoreManager
+import coop.rchain.store.LmdbDirStoreManager._
+import coop.rchain.store.LmdbDirStoreManager
 import coop.rchain.store.KeyValueStoreManager
 
 object RNodeKeyValueStoreManager {
